fix fetching, deleting clients and products, change schema prisma for vat types

+++ backend/src/repositories/client.repository.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateClientInput } from 'src/client/dto/create-client.input';
 import { UpdateClientInput } from 'src/client/dto/update-client.input';
@@ -47,6 +47,14 @@ export class ClientRepository {
   }
 
   async deleteClient(clientId: number) {
+    const invoiceCount = await this.prisma.invoice.count({
+      where: { buyerId: clientId },
+    });
+    if (invoiceCount > 0) {
+      throw new BadRequestException(
+        'Nie można usunąć klienta, ponieważ ma wystawione faktury.',
+      );
+    }
     return this.prisma.client.delete({
       where: { id: clientId },
     });
